feat(config): seprate config & server
