+++ modules/pprof/pprof.go
@@ -16,9 +16,33 @@ import (
 	"github.com/Mrs4s/go-cqhttp/modules/servers"
 )
 
+const pprofDefault = `  # pprof 性能分析服务器, 一般情况下不需要启用.
+  # 如果遇到性能问题请上传报告给开发者处理
+  # 注意: pprof服务不支持中间件、不支持鉴权. 请不要开放到公网
+  - pprof:
+      # pprof服务器监听地址
+      host: 127.0.0.1
+      # pprof服务器监听端口
+      port: 7700
+`
+
+// PprofServer pprof性能分析服务器相关配置
+type PprofServer struct {
+	Disabled bool   `yaml:"disabled"`
+	Host     string `yaml:"host"`
+	Port     int    `yaml:"port"`
+}
+
+func init() {
+	config.AddServer(&config.Server{
+		Brief:   "pprof 性能分析服务器",
+		Default: pprofDefault,
+	})
+}
+
 // runPprof 启动 pprof 性能分析服务器
 func runPprof(_ *coolq.CQBot, node yaml.Node) {
-	var conf config.PprofServer
+	var conf PprofServer
 	switch err := node.Decode(&conf); {
 	case err != nil:
 		log.Warn("读取pprof配置失败 :", err)
